adds heavyarmor trait to vampirelord (#873)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Adds the heavy armor (aka plate training) trait to vampire lords

## Why It's Good For The Game

Currently vampire lord can't even run in the armor it creates at its
pool. This guarantees them the heavy armor trait, so they can actually
make good use of the armor they create.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->