Adds Far Travel Subsystem and Maps in a temporary location. (#866)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

This PR shamelessly ports the FarTravel subsystem from Azure Peak by
https://github.com/[EvenInDeathIStillServe](https://github.com/EvenInDeathIStillServe).
It makes some modifications to better fit Vanderlin's lore and changes
some of the code to fit our requirements better.

Adds a new train sprite by Cannibal Hunter and Aberra to take the place
of the original far travel tiles


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game

Adds in the ability for roles to 'cryo' , which lets them leave the
round and reopen their job slots. This significantly reduces admin
workload and hopefully should reduce any issues that have emerged from
people leaving midround and not informing admins of such.

The only roles excluded from the ability to Cryo are any that are
considered 'Noble Positions' , such as Nobles, Princes, Kings, Hands
etc. This is on request of Aberra, alongside also making sense, as it
doesn't make much sense that a random prince could show up to replace
one that has according to the lore, returned to Kingsfield for an
indeterminate amount of time.

It also does not allow dead people to be cryoed, as this makes very
little sense on Vanderlin, as they are supposed to be buried instead.


![dreamseeker_Lxmg7OxNeo](https://github.com/user-attachments/assets/29a28d4a-d54d-4bb9-90ef-f998d0044c30)
You have to drag your sprite onto the back half of the Train Carriage
for this to work.



<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: tontyGH <39193182+tontyGH@users.noreply.github.com>