NeuFood and Powderspice cleanup and fixes (#931)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->
- FIX rotten egg sprites
- FIX gross soup random messages
- Replaced string css with span functions
- Remove beingeaten proc from bowls
- Replace loc with get_turf(X) for playsound
- Rename powderspice.dm to powder.dm
- Move powder and salt to NeuFood.dm
- Remove comments from both files
- Rename powder to flour
- Give salt and rotten meat new descriptions
- Give specific names for rotten meats (Someone make rotten a flag later
please)
- Remove bowl attack override, it was badly copied and its only purpose
is to stop spam drinking, at the cost of 1 second broken message
forcefeeding and not interacting properly with noble traits (I think)
- Bowl attack by now parent calls when not using a spoon, you only
"drink" with messages when you actually do transfer reagents
- **IMPORTANT** Fixed drug overdoses having two overdose procs but this
leaves these overdoses doing very low damage, needs a maintainer/dev
input on what the value should be currently its `0.25*REM` which makes
it 0.1 a cycle which is pitifully low (the removed 10 a cycle near
instantly kills you however)


## Why It's Good For The Game
Having flour not be flour is strange.

The new descriptions help with what you can use it for flour tells you
its for dough salt tells you its a preservative and rotten food tells
you it can be composted (in a roundabout way).

Every rotten meat being "rotten meat" except eggs was bad.

Everything else is bug fixing or file cleanup,

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->