Court Agents Spawn in the Keep Guest Meeting Room (#804)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

This adds spawnpoints for the court agents within the keep. They are the
three red arrows. Additionally, _"Garrison and Court members know who
you are"_ has been removed from the description for the court agent as
it is untrue.

![image](https://github.com/user-attachments/assets/e20c00a0-32ec-4daa-be6d-df1b3d5638bd)
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->


## Why It's Good For The Game

Currently, court agents do not have a spawnpoint because they are
technically not adventurers. This resulted in them spinning a roulette
wheel for every spawn point in the game, sometimes
[funny](https://github.com/user-attachments/assets/74c19c00-4fa8-4e55-bc2a-656ebae299f6)
and sometimes
[**_not_**](https://github.com/user-attachments/assets/a5a3c5fc-4069-4465-98f4-1189fa5635b5),
occasionally resulting in vandermin intervention.
Instead of the woods like other "adventurers", these spawns have been
placed directly in the keep. This is because an agent's first day is
often spent breaking into the keep or having to reveal their identity to
the guard for entry. From experience, this has resulted in
confrontation, mutilation, or blowing the cover of the agent, as guards
are often investigation targets designated by the Hand.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->