Makes eora acolytes legible to be parents (#901)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

* Makes Eora acolytes non virgins, like all other eora adult church
members are.

## Why It's Good For The Game

Case of "Is this on prupse, or an oversigth" of ss13 code.
So just puting this up to see how the coin will land, 

All Eoran church roles have a check, templers even double check it, to
make sure they are not virgins.
Making them not zizo targets, and legible to be parents.
Kinda strange for just eora acolytes to be an expection.

More interesting & varied restrictions/restrictions expections, creat a
good bases to RP of from.
So, is good.

Discord: NHC

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->