Remove unused skills, change medicine description and give it a dream (#915)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Removes the mathematics, clothesmaking and medicinecraft skills. Could
comment them out instead but thats messy and the githistory is eternal.

Clothesmaking is the only one here that could actually be implemented
with current assets the other ones would need stuff added to give them
something to work with but even still it should be added back in a pr
that makes it do something so it doesn't eat dreams.

On request added the medicine dream from #916

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
You can dream these which do absolutely nothing for you.

closes #640
closes #665
closes #700

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->