Pear Pie's and Grenzelhoft bread spread (#881)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

+ Adds pear pies as fruit pie option
+ Adds crafting recipy for grenzel mett, a meat based toping for bread
from Grenzelhoft

## Why It's Good For The Game

Those are just two foods i was sad about not being able to do in game.
Pear pie is incredibly tasty, why wouldnt you want to make it?
Besides, now pears have more use.

The other thing is a german dish, i got sad about not being able to
make, being that i am german.
So i thougth it be fun to ad another pretty german dish to the
Grenzelhoft culture.

I migth add more food to this if i can be arsed to sprite it.
https://imgur.com/a/mPzbizQ

Discord: NHC

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->