Arrows can be dipped in reagents, cloth and rope can be tied directly to belt slot (#977)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
- Arrows, bolts, and darts can be dipped in any reagent now. Poison
projectile crafting recipes removed. Pyroclastic arrows and bolts
unchanged.
- Cloth and rope can be equipped directly to the belt slot to tie a
belt. Use cloth and rope belts inhand to untie them. Cloth and rope belt
recipes removed.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Gets rid of some slapcrafting for more intuitive gameplay.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->